+++ CHANGES.rst
@@ -19,7 +19,7 @@ CHANGELOG
 
 * Trek detail map does not fit information desks bounds by default
 * Trek detail map does not fit POIs bounds by default
-
+* Increased detail map height on big screens
 
 **BUG fixes**